Add support for cleanIgnore configuration
This adds support for a `cleanIgnore` method/property, which can
receive a string or array of strings indicating names of files or
directories directly under the destination folder to _not_ clean.

This can be used e.g. for using metalsmith for gh-pages while
keeping both the source and build in the repository (akin to how
harp allows it by building from e.g. `./_harp` to `./`).