Added Ignores Support, via Experimental Branch of readdir
This is a concept patch for adding ignores support, based on a yet to be
merged PR for ignores support in readdir.

Long story short, huge directory structures can cause issues with
metalsmith and filtering via plugins can become problematic since
they're not called.

Rather than figure out a way for a plugin to remove thousands of entries
from the files list, this patch allows to simply ignore them from being
generated from the `readdir()` func.

Note that this patch should not be merged in it's current state, since
it relies on a PR not yet merged and pushed to npm.

**!!DO NOT MERGE!!**