Expose the `stats` property in each file metadata
A stat(1) call is made for every single file passed to Metalsmith. The only
field exposed in the metadata was `mode`. This commit exposes the whole stats
object by the `stats` property so that it is quickly accessible without the need
of a new stat(1) call.

The `mode` property is left as is so as to avoid compatibility issue, though the
information is now duplicated and is also accessible via `stats.mode`.

See: https://github.com/segmentio/metalsmith.io/pull/39