Test latest Node version in Travis
Tests Metalsmith with the latest Node version in Travis.

This test is set as an allowed failure, since Node could potentially introduce
breaking changes in new major revisions that we wouldn't want to fail new PRs
on (e.g. something broke in our already existing code), but which we would still
want to be made aware of.