Update to use metalsmith-in-place
Update package dependencies. Remove index.js to generate site (obsolete)
and do so with `metalsmith` CLI and `metalsmith.json`.

Update `src` to use Markdown (not HTML source),
and update links to link to the generated `*.html` files.
Make the links relative so you can preview file: pages in the browser