Solves #361: first implementation of metalsmith.statik
statik with a k to avoid naming collisions with reserved keyword static.
For now the impl only has a setter but it is the goal the 'resolved' static files
would be accessible as an object of Metalsmith.Files, eg { [path]: { stats, contents: Buffer.from(path) }, ...others },
so that their stats can be accessed and their source & destination paths known.
The contents of the file are not loaded, the readonly contents buffer contains the source path as string.